OSIS-5619 - Refactoring
