+++ program_management/ddd/domain/service/has_specific_version_with_greater_end_year.py
@@ -28,7 +28,7 @@ from program_management.ddd.business_types import *
 from program_management.models.education_group_version import EducationGroupVersion
 
 
-class HasVersionWithGreaterEndYear(interface.DomainService):
+class HasSpecificVersionWithGreaterEndYear(interface.DomainService):
 
     @classmethod
     def specific_version_greater_than_standard_year(cls, standard_version: 'ProgramTreeVersion') -> bool:
@@ -38,17 +38,6 @@ class HasVersionWithGreaterEndYear(interface.DomainService):
             offer__acronym=standard_version.entity_id.offer_acronym,
         ).exclude(
             version_name=standard_version.version_name
-        ).exists()
-
-    @classmethod
-    def transition_version_greater_than_standard_year(cls, standard_version: 'ProgramTreeVersion') -> bool:
-        return EducationGroupVersion.objects.filter(
-            Q(root_group__group__end_year__isnull=True) |
-            Q(root_group__group__end_year__year__gte=standard_version.entity_id.year),
-            offer__acronym=standard_version.entity_id.offer_acronym,
-        ).exclude(
-            version_name=standard_version.version_name,
-            transition_name=standard_version.transition_name
         ).exists()
 
     @classmethod
